More progress towards 3.0.  There are no functional changes with this check-in.
git-svn-id: https://regexkit.svn.sourceforge.net/svnroot/regexkit/RegexKitLite@48 83f0d1e4-963b-0410-a2a1-d3bd3f33e299

+++ examples/NSString-HexConversion.m
@@ -18,7 +18,7 @@
     cptr              = (const char *)buffer;
   }
 
-  return(strtol(cptr, NULL, 16));
+  return((NSInteger)strtol(cptr, NULL, 16));
 }
 
 @end
